created adding employee with address(not hierarchical) and salary type id

+++ src/main/java/com/SupermarketPOS/Backend/dto/employee_management/SalaryTypeInput.java
@@ -1,8 +1,9 @@
 package com.SupermarketPOS.Backend.dto.employee_management;
 
 public record SalaryTypeInput(
-        Double basicSalary,
-        Double halfDaySalary,
-        Double overTimeSalary,
-        Double  bonus
+        Float basicSalary,
+        Float halfDaySalary,
+        Float overTimeSalary,
+        Float bonus
+
 ) {}
